% Generated by roxygen2 (4.0.0): do not edit by hand
\name{csv_cat}
\alias{csv_cat}
\title{Concatenate all csv files in a directory into one csv file}
\usage{
csv_cat(dir, outfile)
}
\arguments{
\item{dir}{Name of directory to look for csv files}

\item{outfile}{Filename for concatenated csv file}
}
\value{
A character vector of length one containing the path of the concatenated file
}
\description{
This function uses unix commands to concatenate csv files in a directory into one csv file. Only works if all csv files have
the same columns. All headers except for those in the first file will be stripped
}
\details{
I often produce many csv files from large R analyses as a method for dealing with limited physical memory. e.g.
I output the results of each chunk of an analysis as a csv file, then remove the objects from memory, freeing it for
subsequent analysis. I use this to put these csv files back into one csv. Using unix commands is way faster than R for this.
}

