% Generated by roxygen2 (4.0.0): do not edit by hand
\name{df_list2csv}
\alias{df_list2csv}
\title{Save a list of data.frames as seperate csv's}
\usage{
df_list2csv(df_list, filename, digits = nchar(as.character(length(df_list))))
}
\arguments{
\item{df_list}{A list of data.frame objects}

\item{filename}{The base filename which forms the first part of the csv file's name, followed by the number of the list item}

\item{digits}{The number of digits to coerce the list item number to before appending to filename. The default is
the same number of digits in the length of the list. e.g. 999 items = 3 digits, giving 001, 002, ... 999. This keeps files
sorted in the correct order.}
}
\value{
The base filename for the csv files.
}
\description{
This function takes a list of data.frames and saves them as numbered csv files. Often used in conjunction with
\code{\link{df_split}}.
}

