% Generated by roxygen2 (4.0.0): do not edit by hand
\name{df_split}
\alias{df_split}
\title{Split large data.frame into a list of data.frames}
\usage{
df_split(df, nrows)
}
\arguments{
\item{df}{A data.frame object to split}

\item{nrows}{The number of rows each smaller data.frame should have. If the data.frame cannot be split evenly into that
number of rows, the final data.frame in the list will be smaller than the others.}
}
\value{
A list of data.frames
}
\description{
This function splits an input data.frame into a list of data.frames such that each data.frame in the list has at most
\code{nrows} rows.
}
\details{
I usually use this to split data.frames with many rows into smaller ones which I can save with
\code{\link{df_list2csv}}. This allows me to process each csv individually later to save memory.
}

