% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fast_read}
\alias{fast_read}
\title{Read from a file to a data.frame faster}
\usage{
fast_read(path, sep = ",", header = TRUE)
}
\arguments{
\item{path}{Path to file that you wish to read into a data.frame}

\item{sep}{Seperator character used in file. Default is "," for a .csv file}

\item{header}{Is there a header line in the file?}
}
\value{
a data.frame with the file data
}
\description{
A function to read in data from a file into a data.frame using \code{\link{scan}}, which should be much faster than
\code{\link{read.table}}. This is still not as fast as \code{\link{fread}} from \code{data.table}, but it outputs
a data.frame, so I use when I need a data.frame but the files are big!
}
\details{
This is slower than read.table for small data files, and works much better if there are many rows and few columns.
}

