% Generated by roxygen2 (4.0.0): do not edit by hand
\name{list_to_df}
\alias{list_to_df}
\title{Turn a list into a dplyr 'column-list'}
\usage{
list_to_df(listfordf)
}
\arguments{
\item{listfordf}{A list of R objects which you wish to convert for use in \code{dplyr}. Can be named or unnamed.}
}
\value{
a \code{dplyr} \code{\link[dplyr]{tbl_df}} object
}
\description{
A function to turn a list into a \code{dplyr} 'column-list' which will allow it to be manipulated using \code{dplyr}
functions, particularly \code{\link[dplyr]{summarise}} and \code{\link[dplyr]{do}}.
}
\details{
The idea for this function came from Andrew Macdonald, and was written by Hadley Wickham as a comment on \url{https://gist.github.com/aammd/9ae2f5cce9afd799bafb}
}

